+++ test/recording_studio_root_switchable_test.rb
@@ -58,6 +58,22 @@ class RecordingStudioRootSwitchableTest < Minitest::Test
     refute_includes initializer_source, "config.layout ="
   end
 
+  def test_dummy_accessible_initializer_registers_access_capability_for_recording_studio_v3
+    initializer_path = File.expand_path("dummy/config/initializers/recording_studio_accessible.rb", __dir__)
+    initializer_source = File.read(initializer_path)
+
+    assert_includes initializer_source, "RecordingStudio.register_capability("
+    assert_includes initializer_source, 'child_recordables: [ "RecordingStudio::Access" ]'
+  end
+
+  def test_gemspec_targets_recording_studio_3_and_accessible_0_3
+    gemspec_path = File.expand_path("../recording_studio_root_switchable.gemspec", __dir__)
+    gemspec_source = File.read(gemspec_path)
+
+    assert_includes gemspec_source, 'spec.add_dependency "recording_studio", "~> 3.0"'
+    assert_includes gemspec_source, 'spec.add_dependency "recording_studio_accessible", "~> 0.3"'
+  end
+
   def test_gem_ships_blank_layout_template
     layout_path = File.expand_path("../app/views/layouts/recording_studio_root_switchable/blank.html.erb", __dir__)
     layout_source = File.read(layout_path)